requirements.txt
streamlit>=1.28.0
pandas>=2.1.0
numpy>=1.24.0
tensorflow>=2.15.0
plotly>=5.17.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
twilio>=7.0.0
cryptography>=41.0.0
plotly>=5.17.0
fpdf
fpdf2
pillow